refactor: streamline jsx 에서 js 로 수정
