+++ front/packages/user/src/components/Map.tsx
@@ -18,11 +18,12 @@ import {MeasurePosition} from "@/components/maptool/MeasurePosition.tsx";
 import {MeasureRadius} from "@/components/maptool/MeasureRadius.tsx";
 import SignInfo from "@/components/SignInfo.tsx";
 import LanguageSelector from "@/components/LanguageSelector.tsx";
-import StreamLine from "@/components/StreamLine.jsx";
+import {StreamLine} from "@/modules/streamline/StreamLine";
 
 const Globe = () => {
     const cesiumContainer = useRef<HTMLDivElement>(null);
     useCreateViewer(cesiumContainer);
+    StreamLine();
     return (
         <div id="globe" className={"globe"} ref={cesiumContainer}>
             <header >
@@ -45,7 +46,6 @@ const Globe = () => {
             <SettingTool />
             <CameraInfoDisplay />
             <ObjectToolbox />
-            <StreamLine />
         </div>
     )
 }
